Prototype "Exercise UI" scripts.
DVA Prototype Test Automation / Automate web page
* Use this process to automatically interact with the prototype at
http://localhost:5000/ or http://myservice-disco.herokuapp.com/
to reach certain states.
For example, to start a claim, fill in details and click buttons
to progress through screens.  The purpose of this is to automate
testing the prototype to confirm that recent changes to some feature
have not changed existing behaviour.  And it helps the prototype
developers reach some state in the application from where they
can continue their development from.
* The tool that has been selected for this is Katalon Automation
Recorder, which is based on the Selenium IDE (it hasn’t been
updated since Firefox 55).
* A decent tutorial I have found for it is from DZone -
https://dzone.com/articles/execute-automation-testing-in-3-steps.
* The test scripts are saved in the myservice-prototype/test-scripts
directory of the prototype at https://github.com/AusDVA/myservice-prototype.
* Simply open your web browser of choice (Firefox or Chrome are the
only choices I believe) and install the Katalon Recorder extension.
* Press the Katalon button on the browsers tool bar and its window
should open.
* Next press the Test Suites Open button (the folder) and navigate
to the test-scripts directory.  Open the appropriate test script and
hit play.  This will exercise the UI and get the application in to
some state described in the test script.